Refactor of service architecture
