+++ src/fianchetto_tradebot/common_models/api/quotes/get_options_chain_response.py
