+++ src/fianchetto_tradebot/common_models/api/quotes/get_tradable_response.py
