+++ src/fianchetto_tradebot/server/oex/tactics/incremental_price_delta_execution_tactic.py
@@ -8,7 +8,7 @@ from fianchetto_tradebot.common_models.order.order_type import OrderType
 from fianchetto_tradebot.common_models.order.placed_order import PlacedOrder
 from fianchetto_tradebot.server.oex.tactics.execution_tactic import ExecutionTactic
 from fianchetto_tradebot.server.oex.trade_execution_util import TradeExecutionUtil
-from fianchetto_tradebot.server.quotes import QuotesService
+from fianchetto_tradebot.server.quotes.quotes_service import QuotesService
 
 GAP_REDUCTION_RATIO = 1/3
 DEFAULT_WAIT_SEC = 12
