+++ src/fianchetto_tradebot/server/quotes/etrade/etrade_quotes_service.py
@@ -15,12 +15,12 @@ from fianchetto_tradebot.common_models.finance.option_type import OptionType
 from fianchetto_tradebot.common_models.finance.price import Price
 from fianchetto_tradebot.common_models.finance.priced_option import PricedOption
 from fianchetto_tradebot.common_models.finance.tradable import Tradable
-from fianchetto_tradebot.server.quotes.api.get_option_expire_dates_request import GetOptionExpireDatesRequest
-from fianchetto_tradebot.server.quotes.api.get_option_expire_dates_response import GetOptionExpireDatesResponse
-from fianchetto_tradebot.server.quotes.api.get_options_chain_request import GetOptionsChainRequest
-from fianchetto_tradebot.server.quotes.api.get_options_chain_response import GetOptionsChainResponse
-from fianchetto_tradebot.server.quotes.api.get_tradable_request import GetTradableRequest
-from fianchetto_tradebot.server.quotes.api.get_tradable_response import GetTradableResponse
+from fianchetto_tradebot.common_models.api.quotes.get_option_expire_dates_request import GetOptionExpireDatesRequest
+from fianchetto_tradebot.common_models.api.quotes.get_option_expire_dates_response import GetOptionExpireDatesResponse
+from fianchetto_tradebot.common_models.api.quotes.get_options_chain_request import GetOptionsChainRequest
+from fianchetto_tradebot.common_models.api.quotes.get_options_chain_response import GetOptionsChainResponse
+from fianchetto_tradebot.common_models.api.quotes.get_tradable_request import GetTradableRequest
+from fianchetto_tradebot.common_models.api.quotes.get_tradable_response import GetTradableResponse
 from fianchetto_tradebot.server.quotes.quotes_service import QuotesService
 
 logger = logging.getLogger(__name__)
