+++ src/fianchetto_tradebot/server/quotes/quotes_service.py
@@ -1,12 +1,12 @@
 from fianchetto_tradebot.server.common.api.api_service import ApiService
 from fianchetto_tradebot.server.common.brokerage.connector import Connector
 from fianchetto_tradebot.common_models.finance.option import Option
-from fianchetto_tradebot.server.quotes.api.get_option_expire_dates_request import GetOptionExpireDatesRequest
-from fianchetto_tradebot.server.quotes.api.get_option_expire_dates_response import GetOptionExpireDatesResponse
-from fianchetto_tradebot.server.quotes.api.get_options_chain_request import GetOptionsChainRequest
-from fianchetto_tradebot.server.quotes.api.get_options_chain_response import GetOptionsChainResponse
-from fianchetto_tradebot.server.quotes.api.get_tradable_request import GetTradableRequest
-from fianchetto_tradebot.server.quotes.api.get_tradable_response import GetTradableResponse
+from fianchetto_tradebot.common_models.api.quotes.get_option_expire_dates_request import GetOptionExpireDatesRequest
+from fianchetto_tradebot.common_models.api.quotes.get_option_expire_dates_response import GetOptionExpireDatesResponse
+from fianchetto_tradebot.common_models.api.quotes.get_options_chain_request import GetOptionsChainRequest
+from fianchetto_tradebot.common_models.api.quotes.get_options_chain_response import GetOptionsChainResponse
+from fianchetto_tradebot.common_models.api.quotes.get_tradable_request import GetTradableRequest
+from fianchetto_tradebot.common_models.api.quotes.get_tradable_response import GetTradableResponse
 
 
 class QuotesService(ApiService):
